+++ src/resolvers/user.js
@@ -1,19 +1,39 @@
 import User from '../models/user'
 import bcrypt from 'bcrypt'
 
-import { authenticate } from '../helpers/isAuth'
-
-export const createUser = async (parent, { email, password, role }, { req }) => {
+export const user = async (parent, args, context) => {
   try {
-    authenticate(req)
-    const checkUser = await User.findOne({ email: email.trim() })
+    // if (!context.isAuth) throw Error('Unauthorized')
+
+    const _id = parent ? parent.owner : args._id
+    return await User.findById(_id).lean()
+  } catch (err) {
+    throw err.message
+  }
+}
+
+export const users = async (parent, args, context) => {
+  try {
+    if (!context.isAuth) throw Error('Unauthorized')
+
+    return await User.find({}).lean()
+  } catch (err) {
+    throw err.message
+  }
+}
+
+export const createUser = async (parent, { email, password, role }, context) => {
+  try {
+    if (!context.isAuth) throw Error('Unauthorized')
+
+    const checkUser = await User.findOne({ email: { $regex: email.trim(), $options: 'ig' } })
     if (checkUser) {
-      throw { message: 'User already exists' }
+      throw Error('User already exists')
     }
 
     const hashedPassword = await bcrypt.hash(password.trim(), 15)
     const user = new User({
-      email,
+      email: email.trim(),
       password: hashedPassword,
       role: role ? role : 'user'
     })
@@ -24,12 +44,13 @@ export const createUser = async (parent, { email, password, role }, { req }) => 
   }
 }
 
-export const updateUser = async (parent, { _id, email, password, role }, { req }) => {
+export const updateUser = async (parent, { _id, email, password, role }, context) => {
   try {
-    authenticate(req)
+    if (!context.isAuth) throw Error('Unauthorized')
+
     const checkUser = await User.findById(_id)
     if (!checkUser) {
-      throw { message: 'User not found' }
+      throw Error('User not found')
     }
     let hashedPassword
     let obj = {
@@ -48,12 +69,13 @@ export const updateUser = async (parent, { _id, email, password, role }, { req }
   }
 }
 
-export const deleteUser = async (parent, { _id }, { req }) => {
+export const deleteUser = async (parent, { _id }, context) => {
   try {
-    authenticate(req)
+    if (!context.isAuth) throw Error('Unauthorized')
+
     const checkUser = await User.findById(_id)
     if (!checkUser) {
-      throw { message: 'User not found' }
+      throw Error('User not found')
     }
     await User.findByIdAndDelete(_id)
     return 'User deleted successfully'
